Fix log-handling in Solver-class
Problem:
When using the CLI-version, a (recoverable) error is thrown: In
main.php, the Solver-constructor is called without an argument although
the constructor actually demands a Logger. The variable Solver->log is
theoretically never initialized with anything.

Somehow PHP is able to recover from the error by determining that log
must be WebLogger since it is the only class implementing Logger (even
though WebLogger is in a completely different file).

Solution:
- remove parameter of Solver-constructor
- remove useless if conditions in Solver-constructor and Solver->log-method
- let Solver->log become a new WebLogger()
- move WebLogger class from webfrontend.php to util.php such that
  solver.php can use it (but remove unused __wakeup-method)
- remove unused Logger-interface